json-c updated following move to debian bullsye
The update to json-c - from 0.12.x to 0.15.x - introduced a series of breakages:

 * lh_xxx_table_new() function signature changed, the debug name
   string was dropped

 * the hash key field was switched from "void *" to "const void *"

 * the iterator type is now opaque

 * the error check macro (is_error()) has been removed

In addition various fixes to "silence" warnings generated by the
latest gcc.

Fixes: VRVDR-54768

+++ test/module/src/libvplaned_cstore_cpputest.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2019, AT&T Intellectual Property.
+ * Copyright (c) 2018-2019, 2021 AT&T Intellectual Property.
  * All rights reserved.
  *
  * SPDX-License-Identifier: LGPL-2.1-only
@@ -73,7 +73,7 @@ jsonstr_object_equal(json_object *o1, json_object *o2)
 		return true;
 
 	case json_type_array:
-		int len, i;
+		size_t len, i;
 
 		len = json_object_array_length(o1);
 		if (len != json_object_array_length(o2))
